requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
boto3==1.38.11
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.1.0
markdown==3.5.1
Flask-WTF==1.2.2
